REST_API/requirements.txt
Django
djangorestframework
mlxtend
pandas
request
django-cors-headers